19. Remove Nth Node From End of List
