+++ src/test/kotlin/me/bossm0n5t3r/leetcode/removeNthNodeFromEndOfList/RemoveNthNodeFromEndOfListTest.kt
@@ -1,12 +1,12 @@
-package removeNthNodeFromEndOfList
+package me.bossm0n5t3r.leetcode.removeNthNodeFromEndOfList
 
+import me.bossm0n5t3r.leetcode.utils.ListNode
+import me.bossm0n5t3r.leetcode.utils.ListNodeUtil
 import org.junit.jupiter.api.Test
-import utils.ListNode
-import utils.ListNodeUtil
 import kotlin.test.assertTrue
 
 internal class RemoveNthNodeFromEndOfListTest {
-    private val removeNthNodeFromEndOfList = RemoveNthNodeFromEndOfList()
+    private val removeNthNodeFromEndOfList = RemoveNthNodeFromEndOfList.Solution()
     private val listNodeUtil = ListNodeUtil()
 
     private data class RemoveNthNodeFromEndOfListTestData(
